3/12/2024
- hoàn thiện cơ bản tất cả
- 1 số chức năng có thể thêm: settings để cập nhật top icon.., có thể làm lại trang homepage thêm hotnews, top views, top comment chẳng hạn.

+++ app/Repositories/Listing/ListingRepository.php
@@ -38,9 +38,10 @@ class ListingRepository implements ListingRepositoryInterface
         if (!$listing) {
             return false;
         }
-
         if ($image) {
-            Storage::disk('public')->delete($listing->image);
+            if ($listing->image) {
+                Storage::disk('public')->delete($listing->image);
+            }
             $fields['image'] = Storage::disk('public')->put('images/listing', $image);
         } else {
             $fields['image'] = $listing->image;
